requirements.txt
dj-database-url==0.5.0
Django==3.2.4
psycopg2-binary==2.9.1
python-decouple==3.4
Pillow==9.1.0
django-cors-headers==3.12.0
djangorestframework==3.13.1
gunicorn==20.1.0